+++ changelog.md
@@ -1,10 +1,12 @@
 # 7.8.7
 
+* General AI improvements.
 * Archer fixes and improvements.
   * Fixed archer crossbow.
 * Fixed quotation marks showing up around villager names when migrating worlds from older mod versions.
 * Fixed guards occasionally ignoring active threats.
-* Fixed a bug where guards doing jobs/chores would repeatedly unequip their weapons during combat.
+* Fixed baby, toddler and child villagers not reliably being pickable.
+
 
 # 7.8.6